refactor(neighbor): move mental_state to mind, reserve runtime hook shapes
Two preparatory steps for the upcoming agent.py thin-wrap (PR3):

1. Move src/familiar_agent/mental_state.py to
   src/familiar_neighbor/mind/mental_state.py, leaving a one-line
   re-export shim at the old path so legacy imports still work.  The
   three mind/ siblings that previously imported via the absolute
   familiar_agent.mental_state path (interoception, social_policy,
   appraisal) now use the relative .mental_state import, removing the
   circular shim hop.  mind/__init__.py also re-exports MentalStateBus
   alongside the other top-level cognition engines.

2. Reserve three substrate shapes in familiar_runtime.runtime so PR3
   can wire the embodied behaviours through hooks without changing
   callsites again:

   - RetryDecision dataclass: return-shape for a hook that wants to
     reject a model reply and inject a follow-up user message.
   - InterruptSource Protocol: drain()/empty() polling shim the
     neighbour profile will pass in to surface unprompted user input
     mid-turn (e.g. asyncio.Queue adapter).
   - mid_turn_inject(ctx, iteration) -> list[ContextBlock] hook slot:
     declared on RuntimeHook with a no-op default on RuntimeHookBase so
     concrete hooks can subclass safely today.

AgentRuntime.run_turn now accepts an interrupt_source keyword for
forward-compatibility but does not yet poll it; ReActLoop is unchanged.
PR3 will wire the polling, RetryDecision honouring, and mid_turn_inject
call sites.

Tests:
- test_runtime_hooks.py extended with 4 new cases covering the
  RetryDecision default, the InterruptSource shape, the interrupt_source
  no-poll guarantee, and a tripwire confirming mid_turn_inject is
  declared but not yet invoked.
