+++ src/familiar_neighbor/mind/__init__.py
@@ -9,6 +9,7 @@ from .attention_schema import AttentionSchema
 from .concern_engine import ConcernEngine
 from .default_mode import DefaultModeProcessor
 from .desires import DesireSystem
+from .mental_state import MentalStateBus
 from .meta_monitor import MetaMonitor
 from .prediction import PredictionEngine
 from .relationship import RelationshipTracker
@@ -23,6 +24,7 @@ __all__ = [
     "DefaultModeProcessor",
     "DesireSystem",
     "GlobalWorkspace",
+    "MentalStateBus",
     "MetaMonitor",
     "PredictionEngine",
     "RelationshipTracker",
